Сделал эффект заражения у человечков. Сделал функцию дамага для человечков. Сделал рандомный скину у человечков.

+++ main/game/characters/modules/character_damage.lua
@@ -1,7 +1,7 @@
 -- Атаки персонажей
 local M = {}
 
---Запуск атаки
+--Дамаг по зомби
 function M.damage_zombie(self, message)
 	-- Получили урон
 	local damage = message.damage or 0
@@ -20,6 +20,26 @@ function M.damage_zombie(self, message)
 			position = position,
 			animation_id = hash("effect_zombie_death"), 
 			timer_delete = 3
+		})
+		go.delete()
+	end
+end
+
+-- Дамаг человечка
+function M.damage_human(self, message)
+	-- Получили урон
+	local damage = message.damage or 0
+	self.from_id_object = message.from_id_object 
+	self.live = self.live - damage
+
+	if self.live <= 0 then
+		local position = go.get_position()
+		position.y = position.y + go.get("#body", "size").y / 2
+
+		msg.post(storage_game.map.url_script, "effect", {
+			position = position,
+			animation_id = hash("effect_infection"), 
+			timer_delete = 0
 		})
 		go.delete()
 	end
